requirements.txt
pip>=20.0.0
PySide6
paho-mqtt
cryptography